Refactor the SmapEntry event generation
